Move bughunt selection and bench stats helpers
